requirements.txt
discord.py
aiohttp
colorama
pystyle
requests
flask
flask-socketio
eventlet
